requirements.txt
python-dotenv
langchain
langchain-groq
langchain-community
chromadb
unstructured[docx]
huggingface-hub
sentence-transformers
-e .